feat(comment-and-cleanup): refactor dependencies package
